+++ resources/leiningen/new/luminus/db/src/datetime-deserializers.clj
@@ -1,0 +1,9 @@
+  java.sql.Timestamp
+  (result-set-read-column [v _2 _3]
+    (.toLocalDateTime v))
+  java.sql.Date
+  (result-set-read-column [v _2 _3]
+    (.toLocalDate v))
+  java.sql.Time
+  (result-set-read-column [v _2 _3]
+    (.toLocalTime v))