+++ resources/leiningen/new/luminus/db/src/datetime-serializers.clj
@@ -1,0 +1,15 @@
+  java.util.Date
+  (sql-value [v]
+    (java.sql.Timestamp. (.getTime v)))
+  java.time.LocalTime
+  (sql-value [v]
+    (jt/sql-time v))
+  java.time.LocalDate
+  (sql-value [v]
+    (jt/sql-date v))
+  java.time.LocalDateTime
+  (sql-value [v]
+    (jt/sql-timestamp v))
+  java.time.ZonedDateTime
+  (sql-value [v]
+    (jt/sql-timestamp v))